Move getSuggestedEvents to productions because of path 'productions/suggestion'
